chore: move events related to a stroke's property to a separate file
