+++ Scribble/Services/CanvasStateService/Handlers/PropertyReplayHandler.cs
@@ -1,6 +1,7 @@
 using System;
 using Scribble.Services.CanvasStateService.State;
 using Scribble.Shared.Lib;
+using Scribble.Shared.Lib.Events;
 using Scribble.Utils;
 using SkiaSharp;
