+++ Scribble/Services/CanvasStateService/Handlers/TextReplayHandler.cs
@@ -1,6 +1,7 @@
 using Scribble.Services.CanvasStateService.State;
 using Scribble.Shared.Lib;
 using Scribble.Shared.Lib.CanvasElements.Strokes;
+using Scribble.Shared.Lib.Events;
 using Scribble.Utils;
 using SkiaSharp;
